feat: allow disable link confirmation
There might be site where having a confirmation pop-up is undesired.
Therefore, give the site editors an option to disable the dialog (via
the customizer).

To stay backwards compatible the default value of the newly introduced
"linkConfirmation" control is true, i.e. the confirmation dialog pops up
unless it is disabled explicitly.